Refactor MEM from dos_programs
